+++ src/main/java/com/dao/PicInfoMapper.java
@@ -10,7 +10,7 @@ public interface PicInfoMapper {
 
     int deleteByExample(PicInfoExample example);
 
-    int deleteByPrimaryKey(Byte picId);
+    int deleteByPrimaryKey(Integer picId);
 
     int insert(PicInfo record);
 
@@ -18,7 +18,7 @@ public interface PicInfoMapper {
 
     List<PicInfo> selectByExample(PicInfoExample example);
 
-    PicInfo selectByPrimaryKey(Byte picId);
+    PicInfo selectByPrimaryKey(Integer picId);
 
     int updateByExampleSelective(@Param("record") PicInfo record, @Param("example") PicInfoExample example);
